📝 feat. detail components style module
