+++ src/components/detailPage/commentList/CommentList.tsx
@@ -1,17 +1,17 @@
-import styled from "styled-components";
+import { useState } from "react";
 import Pagination from "@mui/material/Pagination";
 
+import * as t from "./commentList.style";
+import { theme } from "../../../style/theme";
 import Comment from "../comment/Comment";
 import { MainButton } from "../../../elements/Buttons";
 import ReviewModal from "../../modal/reviewModal/ReviewModal";
-import { useState } from "react";
-import { theme } from "../../../style/theme";
 
 const CommentList = () => {
   const [modalOpen, setModalOpen] = useState(false);
 
   return (
-    <MainContainer>
+    <t.MainContainer>
       {modalOpen ? (
         <ReviewModal
           handleClose={() => setModalOpen(false)}
@@ -28,40 +28,16 @@ const CommentList = () => {
       >
         구매평 작성
       </MainButton>
-      <CommentWrapper>
+      <t.CommentWrapper>
         <Comment />
         <Comment />
         <Comment />
-      </CommentWrapper>
-      <Page>
+      </t.CommentWrapper>
+      <t.Page>
         <Pagination count={2} />
-      </Page>
-    </MainContainer>
+      </t.Page>
+    </t.MainContainer>
   );
 };
 
 export default CommentList;
-
-const MainContainer = styled.div`
-  width: 98%;
-  display: flex;
-  flex-direction: column;
-  margin: 20px 0;
-  p {
-    color: ${theme.fc09};
-    font-size: ${theme.fs13};
-    margin: 10px 0;
-  }
-`;
-const CommentWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  border-top: 1px solid ${theme.ls13};
-  margin-top: 30px;
-`;
-const Page = styled.div`
-  margin-top: 30px;
-  display: flex;
-  justify-content: center;
-`;
