+++ src/components/detailPage/deliveryInfo/DeliveryInfo.tsx
@@ -1,11 +1,9 @@
-import styled from "styled-components";
-
-import { theme } from "../../../style/theme";
+import * as t from "./deliveryInfo.style";
 
 const DeliveryInfo = () => {
   return (
-    <MainContainer>
-      <Title>배송정보</Title>
+    <t.MainContainer>
+      <t.Title>배송정보</t.Title>
       <img src="https://cdn.imweb.me/upload/S202001038f7af028ec26d/28bb2a4647f30.jpg" />
       <p>
         - 모든 제품의 배송은 Plastic Free 원칙으로 종이재질로 발송됩니다.
@@ -17,24 +15,8 @@ const DeliveryInfo = () => {
         <br />- 배송이 늦어지거나 일부 제품이 품절인 경우 개별적으로 연락을
         드리겠습니다.
       </p>
-    </MainContainer>
+    </t.MainContainer>
   );
 };
 
 export default DeliveryInfo;
-
-const MainContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  color: ${theme.fc09};
-  font-size: ${theme.fs15};
-  p {
-    line-height: 25px;
-  }
-`;
-const Title = styled.div`
-  padding: 10px 0;
-  box-sizing: border-box;
-  border-bottom: 0.5px solid ${theme.ls08};
-`;
