+++ src/components/detailPage/prodComments/ProdComments.tsx
@@ -1,22 +1,12 @@
-import styled from "styled-components";
-
+import * as t from "./prodComments.style";
 import CommentList from "../commentList/CommentList";
 
 const ProdComments = () => {
   return (
-    <MainContainer>
+    <t.MainContainer>
       <CommentList />
-    </MainContainer>
+    </t.MainContainer>
   );
 };
 
 export default ProdComments;
-
-const MainContainer = styled.div`
-  width: 1200px;
-  display: flex;
-  flex-direction: column;
-  @media (max-width: 1200px) {
-    width: 100%;
-  }
-`;
