+++ src/components/detailPage/prodComments/prodComments.style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const MainContainer = styled.div`
+  width: 1200px;
+  display: flex;
+  flex-direction: column;
+  @media (max-width: 1200px) {
+    width: 100%;
+  }
+`;
