+++ src/components/detailPage/prodDesc/ProdDesc.tsx
@@ -1,41 +1,20 @@
-import styled from "styled-components";
-
-import { theme } from "../../../style/theme";
+import * as t from "./prodDesc.style";
 
 const ProdDesc = () => {
   return (
-    <MainContainer>
+    <t.MainContainer>
       <p>
         더 상쾌한 제로웨이스트 양치시간! 지구샵 스트롱 고체치약으로 시작해요:)
         <br />
         *틴케이스 별도 구매
       </p>
-      <Description>
+      <t.Description>
         <span>원산지</span>&ensp; 대한민국 <br />
         <span>브랜드</span>&ensp; 지구샵 <br />
         <span>배송비</span>&ensp; 전 품목 무료배송 <br />
-      </Description>
-    </MainContainer>
+      </t.Description>
+    </t.MainContainer>
   );
 };
 
 export default ProdDesc;
-
-const MainContainer = styled.div`
-  width: 100%;
-  margin-top: 24px;
-  p {
-    font-size: ${theme.fs15};
-    color: ${theme.fc09};
-    line-height: 1.8;
-  }
-`;
-const Description = styled.div`
-  margin-top: 30px;
-  color: ${theme.fc09};
-  font-size: ${theme.fs13};
-  line-height: 1.8;
-  span {
-    font-weight: 600;
-  }
-`;
