+++ src/components/detailPage/prodDetails/prodDetails.style.ts
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const MainContainer = styled.div`
+  width: 100%;
+  margin: 30px auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  img {
+    width: 100%;
+    object-fit: contain;
+  }
+`;
