+++ src/components/detailPage/prodImg/ProdImg.tsx
@@ -1,9 +1,7 @@
 import { useState } from "react";
-import styled from "styled-components";
 
-interface PropsType {
-  imgUrl: string;
-}
+import * as t from "./prodImg.style";
+
 const ProdImg = () => {
   const [img, setImg] = useState([
     "https://cdn.imweb.me/thumbnail/20220816/ed7c65d28916d.jpg",
@@ -14,12 +12,12 @@ const ProdImg = () => {
     "https://cdn.imweb.me/thumbnail/20220816/ed7c65d28916d.jpg"
   );
   return (
-    <MainContainer>
-      <MainImg imgUrl={url} />
-      <SubImgWrapper>
+    <t.MainContainer>
+      <t.MainImg imgUrl={url} />
+      <t.SubImgWrapper>
         {img.map((item, index) => {
           return (
-            <SubImg
+            <t.SubImg
               key={index}
               src={item}
               onMouseEnter={() => {
@@ -28,34 +26,9 @@ const ProdImg = () => {
             />
           );
         })}
-      </SubImgWrapper>
-    </MainContainer>
+      </t.SubImgWrapper>
+    </t.MainContainer>
   );
 };
 
 export default ProdImg;
-
-const MainContainer = styled.div`
-  width: 49%;
-`;
-const MainImg = styled.div`
-  width: 100%;
-  height: 588px;
-  background-image: url(${(props: PropsType) => props.imgUrl});
-  transition: background-image 0.5s ease-in-out;
-  background-size: cover;
-  @media (min-width: 991px) and (max-width: 1200px) {
-    width: 550px;
-    height: 550px;
-  }
-`;
-const SubImgWrapper = styled.div`
-  display: flex;
-  margin-top: 10px;
-  gap: 3px;
-`;
-const SubImg = styled.img`
-  width: 60px;
-  height: 60px;
-  cursor: pointer;
-`;
